Syncronise kubeconfig command and file output
The kubeconfig file uses different options to what is in the commandline options.
This just syncronises the output to be the same.

In our case we have two clusters each with their own gangway deployment
The both use Dex which uses the same LDAP source for both
so the email we have for a user is the same on both clusters.

Signed-off-by: Pete Brown <pete.brown@powerhrg.com>